# src/main/java/com.elianachv.backend1.proyecto/dto/PacienteDto.java
package com.elianachv.backend1.proyecto.dto;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;

import java.time.LocalDate;
@JsonIgnoreProperties(ignoreUnknown = true)
public class PacienteDto {

    private long dni;

    private String nombre;

    private String apellido;

    private String  domicilioCompleto;

    private LocalDate fechaAlta;

    public PacienteDto() {
    }


    public PacienteDto(long dni, String nombre, String apellido, String domicilio, LocalDate fechaAlta) {
        this.dni = dni;
        this.nombre = nombre;
        this.apellido = apellido;
        this.domicilioCompleto = domicilio;
        this.fechaAlta = fechaAlta;
    }

    public long getDni() {
        return dni;
    }

    public void setDni(long dni) {
        this.dni = dni;
    }

    public String getNombre() {
        return nombre;
    }

    public void setNombre(String nombre) {
        this.nombre = nombre;
    }

    public String getApellido() {
        return apellido;
    }

    public void setApellido(String apellido) {
        this.apellido = apellido;
    }

    public String getDomicilioCompleto() {
        return domicilioCompleto;
    }

    public void setDomicilioCompleto(String domicilio) {
        this.domicilioCompleto = domicilio;
    }

    public LocalDate getFechaAlta() {
        return fechaAlta;
    }

    public void setFechaAlta(LocalDate fechaAlta) {
        this.fechaAlta = fechaAlta;
    }
}